Allow Microsoft.ApplicationInsights to be installed with NuGet.Core.
Source code changes:

https://github.com/mrward/nuget/commit/c003faa756de41f08b66fbaba5d2f06fac78fc17

Fixed bug #30049 - Unable to install Microsoft.ApplicationInsights
https://bugzilla.xamarin.com/show_bug.cgi?id=30049

The Microsoft.ApplicationInsights NuGet package specifies a
minClientVersion of 2.8.50313 which is the version of the NuGet
Package Manager's assembly version released by Microsoft but not the
production version which is 2.8.1. NuGet.Core when built for
MonoDevelop uses the same version for both of these values and does
not match the version generated by Microsoft.

To handle this the 2.8.50313 version is mapped to 2.8.1.